refactor(store): move jsonfile backend into @ontrails/store
