--- connectors/with-jsonfile/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@ontrails/with-jsonfile",
-  "version": "1.0.0-beta.14",
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts",
-    "./package.json": "./package.json"
-  },
-  "scripts": {
-    "build": "tsc -b",
-    "test": "bun test",
-    "typecheck": "tsc --noEmit",
-    "lint": "oxlint ./src",
-    "clean": "rm -rf dist *.tsbuildinfo"
-  },
-  "dependencies": {
-    "@ontrails/core": "workspace:^"
-  },
-  "peerDependencies": {
-    "@ontrails/store": "workspace:^",
-    "zod": "catalog:"
-  }
-}
